+++ tests/unit/test_client_basic_behavior.py
@@ -1,35 +1,15 @@
 import tempfile
-import unittest
 from pathlib import Path
 
+from tests.unit.unit_test_base_class import UnitTestBaseClass
 from tests.utils import run_image_go_nord_client, is_image_empty, are_images_the_same
 
 
-class ClientShould(unittest.TestCase):
+class ClientShould(UnitTestBaseClass):
     # Get blue_suqare.png path from data folder
     data = Path(__file__).parent.parent / 'data'
-    blue_square_path = data / 'blue_square.png'
-    blue_nord_square_path = data / 'blue_nord_square.png'
-
-    def tearDown(self) -> None:
-        # Delete the output image if exists
-        std_output_path = Path('nord.png')
-        if std_output_path.exists():
-            std_output_path.unlink()
-
-    def run_test(self, output_image_path, *args):
-        command, output = run_image_go_nord_client(*args)
-        self.assertFalse(is_image_empty(output_image_path))
-        self.assertTrue(
-            are_images_the_same(self.blue_nord_square_path, output_image_path),
-            f"FAIL: The output image is NOT the same as the expected image \n{command}",
-        )
-        self.assertFalse(
-            are_images_the_same(self.blue_square_path, output_image_path),
-            f"FAIL: The output image IS the same as the expected image \n{command}",
-        )
-
-        return output
+    input_image_path = data / 'blue_square.png'
+    expected_image_path = data / 'blue_nord_square.png'
 
     def test_return_docs_when_given_help_extended_parameter(self):
         _, output = run_image_go_nord_client('--help')
@@ -48,22 +28,22 @@ class ClientShould(unittest.TestCase):
         self.assertEqual('0.1.0', output.strip())
 
     def test_no_output_when_convert_to_nord_palette_and_quiet_short_parameter_provided(self):
-        _, output = run_image_go_nord_client(f'-i={self.blue_square_path}', '-q')
+        _, output = run_image_go_nord_client(f'-i={self.input_image_path}', '-q')
         self.assertEqual('', output)
 
     def test_no_output_when_convert_to_nord_palette_and_quiet_extended_parameter_provided(self):
-        _, output = run_image_go_nord_client(f'--img={self.blue_square_path}', '--quiet')
+        _, output = run_image_go_nord_client(f'--img={self.input_image_path}', '--quiet')
         self.assertEqual('', output)
 
     def test_convert_to_nord_palette_when_given_only_img_in_short_version(self):
         # Check if default file nord.png is created in the current directory
-        run_image_go_nord_client(f'-i={self.blue_square_path}')
+        run_image_go_nord_client(f'-i={self.input_image_path}')
         default_file_path = Path('nord.png')
         self.assertTrue(default_file_path.exists())
 
     def test_convert_to_nord_palette_when_given_only_img_in_extended_version(self):
         # Check if default file nord.png is created in the current directory
-        run_image_go_nord_client(f'--img={self.blue_square_path}')
+        run_image_go_nord_client(f'--img={self.input_image_path}')
         default_file_path = Path('nord.png')
         self.assertTrue(default_file_path.exists())
 
@@ -75,7 +55,7 @@ class ClientShould(unittest.TestCase):
 
             self.run_test(
                 output_image_path,
-                f'--img={self.blue_square_path}',
+                f'--img={self.input_image_path}',
                 f'--out={output_image_path}'
             )
 
@@ -87,6 +67,6 @@ class ClientShould(unittest.TestCase):
 
             self.run_test(
                 output_image_path,
-                f'-i={self.blue_square_path}',
+                f'-i={self.input_image_path}',
                 f'-o={output_image_path}'
             )
